Merging W&D changes from wmaddisn/MesquiteCore into MesquiteProject/MesquiteCore development branch, to allow for new workflow whereby others fork off of MesquiteProject/MesquiteCore.development

+++ Source/mesquite/lib/MesquiteFile.java
@@ -34,6 +34,8 @@ public class MesquiteFile extends Listened implements HNode, Commandable, Listab
 	public static final int LINKED = 0;
 	public static final int INCLUDED = 1;
 	public static final int HOME = -1;
+	
+	public static boolean suppressReadWriteLogging = false;
 
 	private boolean local = true;
 	private URL url;
@@ -980,12 +982,14 @@ public class MesquiteFile extends Listened implements HNode, Commandable, Listab
 			return false;
 		}
 		writing = true;
-		MesquiteModule.mesquiteTrunk.logln( "Saving File: " + fileName);
+		if (!suppressReadWriteLogging)
+			MesquiteModule.mesquiteTrunk.logln( "Saving File: " + fileName);
 		if (directoryName !=null) {
 			writingFileName = directoryName + fileName;
 			backupDirPath = directoryName + "backups" + fileSeparator;
 			backupFileName = backupDirPath + fileName;
-			MesquiteModule.mesquiteTrunk.logln( "In Directory: " + directoryName);
+			if (!suppressReadWriteLogging)
+				MesquiteModule.mesquiteTrunk.logln( "In Directory: " + directoryName);
 		}
 		else {
 			writingFileName = fileName;
@@ -2096,12 +2100,17 @@ public class MesquiteFile extends Listened implements HNode, Commandable, Listab
 
 	/*.................................................................................................................*/
 	public static boolean deleteDirectory(String directoryPath){
+		if (StringUtil.blank(directoryPath))
+			return false;
 		try {
 			File directory = new File(directoryPath);
+			String sep = "";
+			if (!directoryPath.endsWith(MesquiteFile.fileSeparator))
+				sep=MesquiteFile.fileSeparator;
 			if (directory!=null && directory.isDirectory()) {
 				String[] files = directory.list();
 				for (int i=0; i<files.length; i++) {
-					String path = directoryPath + MesquiteFile.fileSeparator + files[i];
+					String path = directoryPath + sep + files[i];
 					File f = new File(path);
 					if (f.isDirectory()) {
 						boolean success = deleteDirectory(path);
@@ -2890,9 +2899,23 @@ public class MesquiteFile extends Listened implements HNode, Commandable, Listab
 		String tempDirectoryName=fdlg.getDirectory();
 		// fdlg.dispose();
 		if (!StringUtil.blank(tempFileName) && !StringUtil.blank(tempDirectoryName)) {
-			MesquiteFile.putFileContents(tempDirectoryName+tempFileName, output, true);
+			MesquiteFile.putFileContents(tempDirectoryName+tempFileName, output,ascii);
 		}
 		MainThread.decrementSuppressWaitWindow();
+	}
+	/*.................................................................................................................*/
+	public static String putFileContentsQueryReturnDirectory(String message, String output){
+		MainThread.incrementSuppressWaitWindow();
+		MesquiteFileDialog fdlg= new MesquiteFileDialog(MesquiteTrunk.mesquiteTrunk.containerOfModule(), message, FileDialog.SAVE);   // Save File dialog box
+		fdlg.setVisible(true);
+		String tempFileName=fdlg.getFile();
+		String tempDirectoryName=fdlg.getDirectory();
+		// fdlg.dispose();
+		if (!StringUtil.blank(tempFileName) && !StringUtil.blank(tempDirectoryName)) {
+			MesquiteFile.putFileContents(tempDirectoryName+tempFileName, output,false);
+		}
+		MainThread.decrementSuppressWaitWindow();
+		return tempDirectoryName;
 	}
 	/*.................................................................................................................*/
 	static boolean w = false;
